Update gitignore to ignore unitypackage meta files
Issue: there are .meta files for .unitypackage files that don't
exist.

Reason: Unitypackage files are compressed unity asset packages.
They are uncompressed once they are imported. A lot of files
Unity manages will have a meta file created for that include Unity
settings for it. The rule of thumb is either a meta file and the
file it is for are tracked together or not at all.

Fix: Need to not track .unitypackage.meta files since we aren't
tracking .unitypackage files. To do this, we need to update the
.gitignore and refresh the index.